+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## UNRELEASE
+
+- **Fix** : Updated order status and state upon transaction failure with notification code 113
+- **Fix** : Modified the cleanPendingOrder cron job to exclude non-HiPay payment methods and restore cancel pending/pending_payment order state
+
 ## 1.27.0
 
 - **Add** : Update OneClick payment